shim: move conn module to a sep file.
